+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. This change
 log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Fixed
+- Focus on the input when the Escape key is pressed. This wasn't working if the mouse was used to click on other components of the UI.
 
 ## [0.19.0] - 2025-05-15
 ### Added